Fix PexprOuterJoinToInnerJoin to transitively convert LOJ to inner joins
The PexprOuterJoinToInnerJoin preprocessing step would not recurse into newly
created children. This would cause only the first LOJ in a tree of LOJs to be
converted to an inner join. This patch fixes the issue by making sure that the
function recurses into its children even if it is newly created.

Co-authored-by: Shreedhar Hardikar <shardikar@pivotal.io>
Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>