Fix cross-checks between unique/primary keys and distribution keys.
We had checks that a unique/primary key index must contain all the
distribution key columns. Otherwise, rows with duplicate index keys might
be stored on different segments, and we would fail to enforce the
uniqueness. But the checks didn't pay any attention to operator classes.
If the index's operator class had a different idea of what "equal" means,
then you might have two rows that are identical according to the index
operator class, but different according to the distribution key hash
operator class, and be stored on different segments.

Modify the checks to also check that the operator classes use the same
equals operator. Refactor so that we can share the code doing the checks
for both CREATE INDEX and ALTER TABLE SET DISTRIBUTED BY.

Fixes https://github.com/greenplum-db/gpdb/issues/6971