Allow exclusion constraints.
Exclusion constraints can be enforced if the table is replicated, or if
the distribution key columns are part of the constraint, with the same =
operator as in the distribution key's hash opclass. This is analogous to
the similar restrictions on unique indexes.

Exclusion constraints on partitioned tables would have similar
restrictions, but because of the way creation of partitioned tables
currently works, checking for those restrictions would be tricky. A
partitioned table is only marked as partitioned after the indexes have
been created, so we cannot check for compatibility with partitioning keys
during index creation, like we can with distribution keys. Refactoring
creation of partitioned tables is more than I can chew right now, so for
now, continue to disallow exclusion constraints on partitioned tables.

Reviewed-by: Zhenghua Lyu <zlv@pivotal.io>
Reviewed-by: Kalen Krempely <kkrempely@pivotal.io>