Report errors for exclusion constraints correctly
When adding unsupported constraints a suitable error message will be
constructed depending on it is an exclusion constraint or not.  There
was a bug that an out-of-array exclusion constraint operator was checked
for that purpose, so the runtime behavior is undetermined.

Fixed by checking the existence of the exclusion constraint operator
array instead of the operator itself.