Forbid DROP PARTITION with no partition name or value.
Previously, this was allowed:

    postgres=# ALTER TABLE parttab DROP PARTITION;
    ALTER TABLE

It dropped the first partition, by rank. In other words, it was a
shorthand for:

    ALTER TABLE parttab DROP PARTITION FOR (RANK (1));

We don't need such a shorthand. It's too much of a footgun; if a user
wants to drop partition, he should name it properly. The documentation
doesn't mention this syntax at all, so no docs changes required.

Reviewed-by: Asim R P <apraveen@pivotal.io>