Dispatched once in one vacuum (#9304)
During the rewriting of vacuum during the 9.0 merge, in commit:
greenplum-db/gpdb-postgres-merge@e26231c the vacuum
workflow was changed to dispatch once per auxiliary table along
with the dispatch for the main table. Before, a vacuum would
dispatch only once and all of the auxiliary tables would be
vacuumed in the same dispatch.

Our heap and ao tables have auxiliary tables. The primary table and
each auxiliary table are dispatched separately. One vacuum will have
too many dispatches, which will cause performance problems.

In postgres, we only verify the permition on the main table. But if
we dispatch auxiliary tables separately, the segment will verify
the permition on auxiliary tables. This will cause failure in
permission verification.

We refactored the code so that each vacuum is dispatched only once.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>