Refactor SliceTable's slices list into an array.
Store the slices as an array, instead of List, for faster and more
convenient access to individual slices.

Rename Slice struct to ExecSlice. I'm about to change things so that we
build a representation of the slice table already at planning time, so
let's make it clear that this is the executor's version of the slice
table.

Remove the Node "type" tag from the ExecSlice struct. We never pass them
around separately, they're always part of the array in SliceTable.

Reviewed-by: Taylor Vesely <tvesely@pivotal.io>