Remove redundant gangSize and direct dispatch info from ExecSlice.
We were using two different representations for the list of segments to
run a slice on. ExecSlice->directDispatch.contentIds was used when the
slice was direct-dispatched, and ExecSlice->segments was used otherwise.
We don't need that distinction.

ExecSlice->gangSize was always set to the length of the segments list. We
don't need a separate field for it, we can just look at the length of the
segments list directly.

Reviewed-by: Taylor Vesely <tvesely@pivotal.io>