Don't set 'numsegments' on General, Entry or OuterQuery locus.
It's not clear what 'numsegments' means on a General, Entry or OuterQuery
locus. General means that the path is executable anywhere, and Entry means
it's executable in the dispatcher node, and neither of those really have
sensible concept of 'numsegments'. And OuterQuery means "whatever locus
the outer query has". So remove the 'numsegments' argument from their
constructors, set it to -1 instead. This requires adjusting a few
assertions that assumed that numsegments is set to something that looks
valid for those locus types.

Reviewed-by: Pengzhou Tang <ptang@pivotal.io>
Reviewed-by: Ning Yu <nyu@pivotal.io>