Simplify selectPartition().
Simplify selectPartition(), by using the B-tree opclass's comparison
function, instead of separate <, =, and > functions

This has no user-visible effect, except for one corner case: If you create
a custom b-tree operator class that is incomplete, so that it has the
equality operator but is missing < > and such, you could previously still
use it as list partitioning key. Now you get an error when you try to
insert to the table. This is user error; don't create an incomplete
operator class. On PostgreSQL v10 partitioning, you get an error already at
CREATE TABLE. We had a test case for that in the 'gporca' test, move it to
'partition', since it doesn't have much to with ORCA. However, pg_upgrade
currently cannot handle such a table, so drop the table before running
pg_upgrade (it was dropped before too, as the whole 'orca' schema is
dropped at the end of the test)

Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>