Only include server-build resources for prod pipelines
The pipeline defined resources for server-build artifacts based on the
OS types being configured; however, these resources were only utilized
in production pipelines. This resulted in an error when setting a dev
pipeline since the resource was unused. This commit only defines the
resources if we are configuring a production pipeline.

Authored-by: Bradford D. Boyle <bboyle@pivotal.io>