Fix node out/read routines for Partition and PartitionRule structs.
The functions for PartitionRule in outfuncs.c and readfuncs.c were missing
the 'parisdefault' field. That was harmless, because those functions are
in fact unused; we never materialize PartitionRule structs to disk.

To fix, add the field, and to reduce the chance of similar bugs in the
future and to reduce the amount of duplication, remove binary versions of
these functions from outfast.c/readfast.c.

Also remove the binary versions of _readPartition() and _outPartition()
while we're at it. No fields were missing from those, but less duplicated
code is good.

Discussion: https://github.com/greenplum-db/gpdb/issues/9390
Reviewed-by: Georgios Kokolatos <gkokolatos@pivotal.io>