Revert create pathkey in `convert_subquery_pathkeys` and push down gather motion
In upstream, it does not create a new pathkey in convert_subquery_pathkeys function. It also raises an issue in gpdb, so revert it.

After that, we need to handle a common case that a gather motion should keep data order however subqueryscan can not provide.

e.g. create a view with an ORDER BY:
```CREATE VIEW v AS SELECT va, vn FROM sourcetable ORDER BY vn;```
and query it:
```SELECT va FROM v_sourcetable;```

In the planer, a subqueryscan is created at the top of the view. In upstream, it is fine, even it does not have a path key, the data is physically sorted well. But for us, we have a Gather Motion upon subqueryscan. The data is sorted on each segment, later Gather Motion has no indication of how to merge tuple.

To deal with this problem, just push down the gather motion under a subqueryscan, if a subqueryscan does not have path key but its subpath has.

fix issue: #8987