Implementing multi-phase grouping sets. (#9219)
This is a work based on multi-phase aggregation in master. The idea is
to perform grouping sets aggregation in partial phase and then perform
normal aggregation in final phase.

In partial phase, we attach a GROUPINGSET_ID to each transvalue. In
final phase, we include this GROUPINGSET_ID at the head of sort keys and
group keys for group aggregation. This will ensure correctness in the
case of NULLs.