Remove obsolete optimized code from EncodeDateTime.
We had optimized this piece of code in Greenplum, but in PostgreSQL 9.6,
commit aa2387e2fd, PostgreSQL made a similar optimization. In the 9.6
merge, I picked up the PostgreSQL version, but left the Greenplum version
in a commented out block, with the plan to do some performance testing to
see if we can switch to the PostgreSQL version now.

I did that performance testing now, and it seems that the old GPDB version
was about the same speed as the new PostgreSQL version. I used this to
test it:

    -- generate test data
    create table tstest  (distkey int4, ts timestamp) distributed by (distkey);
    insert into tstest
      select 1, g from generate_series(now(), now()+ '1 year', '1 second') g;
    vacuum tstest;

    -- test query
    \timing on
    select min(ts::text) from tstest;

The query took about 15 seconds on my laptop, and 'perf' says that about
10% of the CPU time was spent in EncodeDateTime, with both versions.

Reviewed-by: Georgios Kokolatos <gkokolatos@pivotal.io>