Fix pushdown of joins and aggregates to foreign servers.
The planner code assumed that a foreign path is only created for a
base relation. But that is not a valid assumption, a FDW may push down
joins and aggregates, too, which will be represented by the join or
upper rels in the planner. The code in create_foreign_path() tried
to look up the EXECUTE ON attribute from the RelOptInfo->ftEntry field,
but that was only set for base relations, which caused a crash when
pushing down a join or an aggregate.

To fix, propagate the 'exec_location' field to join and upper rels, just
like the fdwroutine and serverid fields are propagated. Unfortunately this
creates a rather large diff, those fields are set in many places. That has
the potential for bugs-of-omission, as we merge with upstream, but I don't
see a better way. If such a bug happens, though, I believe the consequence
is just a graceful error from the planner, or a missed push-down, so I
think we can live with it.

Fixes github issue https://github.com/greenplum-db/gpdb/issues/9209

Reviewed-by: Adam Lee <ali@pivotal.io>