Fix row-count estimate for the lower stages of DQA plans.
In a query with a single DISTINCT-qualified aggregate, the row count
estimate for the bottom deduplication aggregate steps was taken from the
overall aggregation's row count estimate. That could be dramatically
different. For example, in this query from the regression tests:

> explain select sum(distinct b) from olap_test_single;
>                                                   QUERY PLAN
> ---------------------------------------------------------------------------------------------------------------
>  Finalize Aggregate  (cost=166.23..166.24 rows=1 width=8)
>    ->  Gather Motion 3:1  (slice1; segments: 3)  (cost=166.19..166.22 rows=1 width=8)
>          ->  Partial Aggregate  (cost=166.19..166.20 rows=1 width=8)
>                ->  HashAggregate  (cost=166.16..166.19 rows=1 width=4)
>                      Group Key: b
>                      ->  Redistribute Motion 3:3  (slice2; segments: 3)  (cost=165.00..165.99 rows=11 width=4)
>                            Hash Key: b
>                            ->  Streaming HashAggregate  (cost=165.00..165.33 rows=11 width=4)
>                                  Group Key: b
>                                  ->  Seq Scan on olap_test_single  (cost=0.00..115.00 rows=3334 width=4)
>  Optimizer: Postgres query optimizer
> (11 rows)

Before this patch, the Streaming HashAggregate at the bottom had a row count
estimate of 1 rows. 1 row is correct for the overall query, as an aggregate
query with no GROUP BY always returns one row, but wildly incorrect for the
deduplicating Streaming HashAggregate. It returns as many rows as there are
distinct values, the aggregation that rolls them up to one row only happens
in the Partial and Finalize Aggregate steps.

Reviewed-by: Taylor Vesely <tvesely@pivotal.io>