Remove Alien memory account.
When executor nodes are initialized at executor startup, in the
ExecInitPlan() stage, any nodes that are not going to be executed in the
current slice were assigned to so-called Alien memory account. Previously,
that was done to keep the useless nodes out of the "real" memory balances,
but nowadays we normally don't bother initializing alien nodes in the
first place. Alien node elimination can be disabled with 'set
execute_pruned_plan=off', but that's a developer option that people
shouldn't normally mess with. So in normal operation, the Alien memory
account is never used.

The Alien memory account was kept around when the alien node elimination
was implemented (see commit 9b8f5c0b6e). The idea was that we could turn
it on/off, and see how much we're saving by looking at the memory usage
in the Alien memory account when it's turned 'off'. But that's hardly
interesting anymore, we know that alien node elimination is useful, and
it has worked great in production for some time now.

We could probably get rid of the 'execute_pruned_plan' GUC altogether at
this point, but I kept it for now. If you do turn it off, all the alien
nodes will now get their own memory accounts, like non-alien nodes.

Reviewed-by: Venkatesh Raghavan <vraghavan@pivotal.io>