Implement PXF Foreign Data Wrapper
PXF, (Platform Extension Framework) provides access to external data in
Greenplum[1]. Previously, it was based on external tables, this commit
introduces the FDW that can be used to communicate with PXF.

* Provide the skeleton for PXF Foreign Data Wrappers.

* Add pxf_fdw_version()

* Validation for WRAPPER, SERVER, USER-MAPPING, and FOREIGN TABLE options

* Only build pxf_fdw when --enable_pxf is present PXF is enabled by default,
but this gives the user the option of turning off the building of the PXF
contrib module with --disable-pxf.

* Integrate FDW with PXF's fragmenter and bridge
This allows a Greenplum user to create foreign tables that read
from external data via PXF. See [2] for documentation.
The only wire format (format used to communicate between PXF JVM and
Greenplum segments) supported is going to be TEXT for FDW. Other formats
which come across as a binary stream of data (like Parquet) are not
implemented. We validate and pass along the following options:

	- header
	- delimiter
	- quote
	- escape
	- null
	- encoding
	- newline
	- fill_missing_fields
	- force_not_null
	- force_null
	- reject_limit
	- reject_limit_type (percent/rows)

and enforce precedence rules for `other_options`.
For options other than protocol, resource, format, reject_limit*, use
the following precedence rules:

	- Table options take precedence over all other options
	- User-Mapping options take precedence over server and wrapper options
	- Server options take precedence over wrapper options

* Add support for and validation of log_errors.

* Complete FDW write (master only)

This is facilitated by externalizing a new function in COPY: `BeginCopy()`.

* Introduce a `config` option for SERVER:
Access to Hadoop clusters is nuanced, because with a single set of configurations
users are able to access HDFS, Hive, HBase and other services.
Suppose an enterprise user has a Hortonworks hadoop installation that includes
HDFS, Hive, and HBase. We would configure one server per technology we access,
for example:

     CREATE SERVER hdfs_hdp
          FOREIGN DATA WRAPPER hdfs_pxf_fdw
          OPTIONS ( config 'hdp_1' );

     CREATE SERVER hive_hdp
          FOREIGN DATA WRAPPER hive_pxf_fdw
          OPTIONS ( config 'hdp_1' );

     CREATE SERVER hbase_hdp
          FOREIGN DATA WRAPPER hbase_pxf_fdw
          OPTIONS ( config 'hdp_1' );

To reduce the amount of configuration required for each server, we introduce a
new option `config`. This new option provides the name of the server directory where
the configuration files reside. In the example above, configuration files live
in the `$PXF_CONF/servers/hdp_1` directory, and all three servers share the same
configuration directory.

* Default wire_format to CSV

The wire_format in PXF defaults to CSV. Only when the file format is
tab-delimited text, we will use TEXT as the wire_format. This commit
makes CSV the default wire_format for PXF

* Add column projection

* Pass filter string (WHERE clauses) to PXF

[1] https://github.com/greenplum-db/pxf
[2] https://github.com/greenplum-db/pxf/blob/pxf-fdw-d/PXF_FDW.md

Co-authored-by: Oliver Albertini <oalbertini@pivotal.io>
Co-authored-by: Raymond Yin <ryin@pivotal.io>
Co-authored-by: Francisco Guerrero <aguerrero@pivotal.io>