Resurrect constraint checking on external tables.
I initially thought that this was dead code, because you can't create a
CHECK constraint on an external table normally. However, when you exchange
an external table with a table partition, the partition's CHECK
constraints, which check the partition boundaries, are applied to the
external table.

This fixes a regression failure in 'partindex_test' test.

Resurrect the old partition checking code, so that you get the same
behavior as before. I'm not convinced this is really the best behavior,
but this lets us move forward while we discuss what behavior we actually
want.

Discussion: https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/v-ZTreV0ud4/gZutaKYNDQAJ