Clarify what EState->currentSliceId means.
Clarify the comment to explain how it's set during initialization, and
how it's set during execution. Also, stop modifying it in ExecEndMotion().
As far as I can tell, it was harmless, but also confusing, as ending the
motion doesn't really change the slice we're executing. There was no
corresponding line to set it when the motion was started.

Reviewed-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>