Remove Plan->motionNode field.
It was used for two things: to determine which nodes are "alien", in
ExecInitNode(), and to check if a slice is sending to another slice, in
ShouldPrefetchJoinQual(). The first use was removed in commit 2bc04a199f,
and the second in the previous commit 36a93ba806. So it was unused now.

This gets rid of the step in executor startup that walked through the plan
tree, and labeled every plan node with the parent Motion node. Modifying
the motionNode field in the Plan tree in an already-built plan tree
always felt iffy.

Reviewed-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>