Remove most planner GUCs from PlannerConfig.
There's no need to carry around GUCs in this structs that are never
changed in the planner. That's most of the GUCs. Only a few are
temporarily disabled, when recursing into a subquery; keep those. (I'm not
convinced that the PlannerConfig mechanism is a good way to disable those
things either, but I'll leave refactoring that for another day.)

Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>