Remove RowTrigger executor node type, used for firing triggers with ORCA.
If a table has triggers, ORCA generated a plan with a special RowTrigger
node to fire them. That's different from how triggers are normally fired,
but I don't see any good reason for doing things differently in
ORCA-generated plans. Remove RowTrigger node type from the backend, and
hide all triggers from ORCA so that it doesn't generate them anymore.
Triggers are now fired the usual way for ORCA-generated plans.

This feature was disabled by default, and ORCA would fall back for any
DML on tables with triggers, but it could be enable by setting
optimizer_enable_dml_triggers=on. However, it had been broken for a
while. To tet it, I tried running the 'triggers' regression test like
this:

    PGOPTIONS="-coptimizer_enable_dml=on -coptimizer=on" ./pg_regress --init-file=init_file create_function_1 triggers

I tried that on master, 6X_STABLE and 5X_STABLE, and it segfaulted or
failed an assertion all three.

This commit removes the optimizer_enable_dml_triggers GUC. I don't see
any need for such a flag anymore, now that the triggers are fired in the
same way with or without ORCA. It was undocumented, anyway.

UPDATE on a table with triggers still falls back to the Postgres planner,
because the executor isn't prepared to fire UPDATE triggres on a split
Update. The Postgres planner has the same limitation, but it affects ORCA
more because ORCA plans all UPDATEs as split Updates, whereas the Postgres
planner uses a SplitUpdate only when the distribution key columns are
modified.

Reviewed-by: Venkatesh Raghavan <vraghavan@pivotal.io>