Remove feature: password_hash_algorithm=SHA-256
It didn't really provide any security over the default md5 method. And in
PostgreSQL v10, we'll get SCRAM authentication, which is stronger. To
reduce conflicts with the PostgreSQL v10 merge, remove this GPDB-specific
feature now.

This also adds a check to pg_upgrade. I opted to refuse the upgrade if
there are any SHA-256 hashed passwords. Alternatively, we could clear the
passwords after upgrade, but it seems better to give the user a heads up.

Reviewed-by: Daniel Gustafsson <dgustafsson@pivotal.io>