Use zstd for wal_compression
With postgres 9.5 merge GPDB got the capability to compress the full
page images or backup blocks in wal. But given pg_lz is not great and
zstd is better compression algorithm, use zstd instead of pg_lz to
compress the FPI in GPDB.

Just to give a data point, 6000MB of total wal is created on each
segment for ICG. With wal compression enabled, approximately by 700MB
wal size gets reduced for each segment.

Reviewed-by: Adam Lee <ali@pivotal.io>
Reviewed-by: Jesse Zhang <jzhang@pivotal.io>
Reviewed-by: Paul Guo <paulguo@gmail.com>
Reviewed-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>