Remove ORCA expected output for file_fdw, use an "init_file" instead
The only difference in the ORCA-specific expected output was the
"optimizer=on" in the EXPLAIN outputs. Use an init_file to mask that
difference, instead.

I think this might fix the error that @bzhaoopenstack reported at
https://github.com/greenplum-db/gpdb/issues/9533#issuecomment-586899241,
although I'm not entirely sure why he's observing the difference in the
first place. But it's a good idea to avoid unnecessary alternative
expected output files in any case.