Refactor sorted Motion's merge heap.
Instead of having CdbTupleHeapInfo structs to hold the tuples currently
in the heap, store each tuple in a TupleTableSlot. That's what similar
code in nodeMergeAppend.c does.

I did some quick testing of this on laptop, and didn't see much of
difference. But I think this is more clear, and in the long term, I'd
like to get rid of GenericTuples, and this is a step in that direction.

This inlines the CdbMergeComparatorContext struct directly into
MotionState. Having a separate struct made some sense when we still used
to have two implementations of the heap, but that was removed in commit
726fb52a13.

Reviewed-by: Adam Lee <ali@pivotal.io>