Resolve FIXME at Sort node creation.
The 'add_keys_to_targetlist' argument is a GPDB addition. The upstream
behavior is as if add_keys_to_targetlist=true. The 'false' option was
added in GPDB, and was formerly used in the GPDB grouping sets planning
code, which doesn't exist anymore. There's no reason to believe that
we should do differently from upstream in create_sort_plan(), so change
the call to pass 'true'.

Reviewed-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>