Materialize before refering outerParams to avoid rescan of motion.
Commit 6257972 can avoid PANIC for lateral join when inner plan
contains a limit clause but it does not finish all the job. If
the outer plan produces more than one tuple, it will lead to
rescan of motion node which is not supported in Greenplum.

This commit fixes the issue by make a material plan node above
the gather motion at very early stage before the refering of outer
params.