Remove unused things from initdb
We had over time accumulated a few unused remains of hacks gone past.
Versionchecking is no longer done with a string comparison on a hard-
coded string, and the mirror creation support was already half-removed
with only dead code remaining (which in turn caused quite a diff to
upstream for no benefit). Remove it all.

Reviewed-by: Jacob Champion <pchampion@pivotal.io>
Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>