Align initdb a little with upstream
This rearranges, and puts back, some code to align initdb with upstream
in cases where that can be done safely. One case of exit_nicely() was
decorated with break; in a switch statement, and that has been replaced
by exit(1) instead as the break served the purpose of avoiding compiler
warnings on fallthrough. This aligns us with what is happening in the
merge branch, as commit 3913a40ff104fd26f795c3d4c402ec27a1f86385 will
remove exit_nicely altogether.

Reviewed-by: Jacob Champion <pchampion@pivotal.io>
Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>