Fix data reorganization when gp_use_legacy_hashops is true
The ALTER TABLE command reorganizes the data by using a temporary table,
if a "distributed by" clause is specified without the opclasses, the
default opclasses will be chosen.  There was a bug that the non-legacy
opclasses are always chosen, regarding the setting of
gp_use_legacy_hashops.  However the table's new opclasses are determined
with gp_use_legacy_hashops, so when gp_use_legacy_hashops is true the
data will be incorrectly redistributed.

The issue also exists in CTAS.

Fixed by choosing the default opclasses for the temporary table
according to gp_use_legacy_hashops.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>
Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>
Reviewed-by: Shreedhar Hardikar <shardikar@pivotal.io>