Don't bother inlining toasted datums when converting an ancient AO table.
As far as I can see, the toasted datums should be fine, and there's no need
to inline them. I believe we were passing 'true' out of sheer paranoia. But
I want to change this now, because I'm about to refactor
memtuple_form_to(), removing the code to inline toasted datums, and this is
one of only two callers that was passing 'true'.

Unfortunately I have no means of testing this. In order to hit this
codepath, you need to binary-upgrade a cluster that was originally created
with GPDB 4.1 or below (I'm not sure about the version number, but
something from that era). This is currently dead code, as pg_upgrade
doesn't currently support upgrading from a 4.X cluster anyway, but maybe
that will happen some day.

Reviewed-by: David Kimura <dkimura@pivotal.io>