Revert unnecessary changes vs upstream in 'inherit' test.
The extra 'dummy' column doesn't server any useful purpose. It was
introduced back in 2006, with a commit message "fix some tests". Maybe
it was needed back then to make the test pass on GPDB, perhaps because
gpdiff was less advanced back then, but in any case, no need to carry
this diff vs. upstream nowadays.