Fix execution of functions with EXECUTE ON options in utility mode.
It's not entirely clear what the behavior should be in utility mode.
With this commit, the EXECUTE ON options are effectively ignored in
utility mode, so the function is executed locally on the node you're
connected to. To achieve that, force Entry locus for the FunctionScan
path in the planner, when in utility mode. This is how other kinds
of scans are planned in utility mode, too.

Without this fix, the join queries added in the test case, with
EXECUTE ON SEGMENTS function, generated a plan with a Motion, even in
utility mode, which failed an assertion in the executor. EXECUTE ON
INITPLAN had similar issues, causing an error in the executor, when it
tried to open a non-existent tuplestore file.