Fix expected modcount values in behave tests
Now that we no longer maintain modcount for append-optimized tables on
master, analyzedb was changed to obtain sum of modcounts from segments
in 67321b1ceebc6d.  This commit adjusts modcounts expected by tests
accordingly.  E.g. expected modcount of 2 needs to be 6 now (sum of
modcount from 3 segments).  In another test, expected modcount of 2
now becomes 3 (1 on seg0 + 2 on seg1 + 0 on seg2).

Note that expecting a modcount to have a specific value makes the
tests depend on demo cluster configuration.  That dependency already
exists in some ICW tests, so it seems alright.