Simplify the implementation of "dedup" semi-join paths.
There are two parts to this commit:

1. Instead of using ctids and "fake" ctids to disambiguate rows that have
   been duplicated by a Broadcast motion, generate a synthetic 64-bit
   rowid specifically for that purpose, just below the Broadcast. The old
   method had to generate subplan IDs, fake ctids etc. to have a
   combination of columns that formed the unique ID, and that entailed a
   lot of infrastructure, like the whole concept of "pseudo columns". The
   new method is much simpler, and it works the same way regardless of
   what's below the Broadcast. The new 64-bit rowids are generated by
   a new RowIdExpr expression node. The high 16 bits of the rowid are
   the segment ID, and the low bits are generated by a simple counter.

2. Replace the cdbpath_dedup_fixup() post-processing step on the Plan
   tree, by adding the unique ID column (now a RowIdExpr) in the Path's
   pathtarget in the first place. cdbpath_motion_for_join() is the
   function that does that now. Seems like a logical place, since that's
   where any needed Motions are added on top of the join inputs. The
   responsibility for creating the Unique path is moved from
   cdb_add_join_path() to the create_nestloop/hashjoin/mergejoin_path()
   functions.

This slightly changes the plan when one side of the join has Segment or
SegmentGeneral locus. For example:

select * from srf() r(a) where r.a in (select t.a/10 from tab t);

With this commit, that's implemented by running the FunctionScan and
generating the row IDs on only one node, and redistributing/broadcasting
the result. This is not exactly the same as before this PR: previously, if
the function was stable or immutable, we assumed that a SRF returns rows
in the same order on all nodes and relied on that when generating the
"synthetic ctid" on each row. We no longer assume that, but force the
function to run on a single node instead. That seems acceptable, this kind
of plans are helpful when the other side of the join is much larger, so
redistributing the smaller side is probably still OK.

This also uses the same strategy when the inner side is a replicated table
rather than a SRF. That was not supported before, but there was a TODO
comment that hinted at the possibility.

Fixes https://github.com/greenplum-db/gpdb/issues/9741, and adds tests
for it.

Reviewed-by: Zhenghua Lyu <zlv@pivotal.io>