Make fault injector configurable (#9532)
Definition FAULT_INJECTOR is hardcoded in a header(pg_config_manual.h) file.
Fault injector is useful, but it may introduce some issues in production
stage, like runtime cost, security problems. It's better to enable this
feature in development and disable it in release.

To achieve this target, we add a configure option to make fault injector
configurable. When fault injector is disabled, some tests using this feature
should be avoided to run ICW. Under isolation2 and regress, there are a lot of
tests. Now, all tests under isolation2 and regress that depend on fault injector
are moved to a new schedule file. The pattern name of it is XXX_faultinjector_schedule.

**NOTE**
All tests that depend on fault injector are saved to the XXX_faultinjector_schedule.
With this rule, we only run tests that don't depend on fault injector when fault injector
is disabled.

The schedule files used for fault injector are:
src/test/regress/greenplum_faultinjector_schedule
src/test/isolation2/isolation2_faultinjector_schedule
src/test/isolation2/isolation2_resgroup_faultinjector_schedule

Reviewed-by: Asim R P <apraveen@pivotal.io>
Reviewed-by: Ashwin Agrawal <aagrawal@pivotal.io>