gp_execution_segment is not a set-returning function.
I don't understand why it was marked as such in the catalogs. By accident?
It's been like that since the very first commit in the historic GPDB
repository that introduced it back in 2007. I'm amazed that it worked at
all, but clearly it doesn't really return a set, just a single integer.

I spotted this while working on the PostgreSQL v12 merge. Newer versions
of PostgreSQL are more strict on whichg parts of a query a set-returning
function can appear, and some regression tests were failing because of
this.

Reviewed-by: Jimmy Yih <jyih@pivotal.io>