Refactor url parsing for external tables
The previous coding implemented a two stage API where the make_url function
was first called with a NULL buffer to calculate the required length of the
buffer, and then called again with a correctly sized buffer.  This enabled
safe usage of the string copying as long as the API was followed, but was
prone to overflows in case a too small buffer was passed in as the length
wasn't a parameter to make_url.

This refactors make_url to instead build the URL without a second call,
using a dynamic buffer. Rather than using a pass-by-reference buffer it
now returns an allocated and guaranteed-to-be terminated buffer. This makes
the processing a hair more effective, and removes compiler warnings from
unsafe string operations.

Also convert an errorcheck into an Assertion as the error condition cannot
hold. The error handling was this:

	if ((strlen(url) - strlen(hostname) + strlen(hostip)) < 0)
		error..

hostname is copied from url using an offset and length (where the offset is
non-zero), and hostip is guaranteed to be a string of at least length 0.

Reviewed-by: Heikki Linnakangas <hlinnakangas@pivotal.io>