Remove incorrect function prototypes and unnecessary headers.
As @hidva pointed out, the prototype for pg_plan_query() in gpdbdefs.h
was incorrect. On closer inspection, the preprocess_query_optimizer()
prototype at least was also incorrect. But none of these prototypes in
gpdbdefs.h are actually needed. Some of them are not used in ORCA at all,
and others in the headers that are #included. Remove them all.

Also remove a few of the #includes that are not needed. Many more of them
could probably be removed, I didn't systematically go through all of them,
I just removed a few that seemed unlikely to be needed.