Fix wait policy for SELECT ... FOR ... NOWAIT (#9726)
The wait policy, NOWAIT or SKIP LOCKED, only affects how SELCT locks
rows as they are obtained from the table. They are used by the execution.
The SELECT statement with locking clause may upgrade the lock mode.
This PR removes the nowait logic in opening the relation in the parser analyze
stage. So, the SELECT statement with locking clause will wait for the table
lock if it can't hold the table lock.

Co-authored-by: Zhenghua Lyu zlv@pivotal.io