Partial revert Sharedsnapshot refactor PR (#9845)
* Revert "Refactor Sharedsnapshot"

This reverts commit cfc911d846fa5bf6bfa0b70eee406ae3c98460ca.

we create a DSM for sharedsnapshot synchronization between writer and reader. the DSM lifecycle is the same as a session, which we use dsm_ping_maping to mark dsm_segment.

However, in centos 6, the DSM will randomly be free after some test cases even the session does not exist. It seems to related to dsm itself.

Partially revert this part of the code. still left shared snapshot dump to DSM.

* Dump sharedsnapshot to DSM

Cursor writer gang dump snapshot using dsm instead of BufFile. A loop
buffer is kept in Sharedsnapshot to store dump dsm handle. Each cursor
declaration, gpdb dispatch to QE twice. The first dispatch, we force writer
gang dumping snapshot. The Second dispatch, we create a new reader gang to
execute cursor fetch and do not wait for reader gang finishes. Writer gang does
not know when reader gang sync the snapshot done, but must be in a short
period of time after the first dispatch done. The loop buffer is big enough,
so we can repeatedly using this space to store snapshot dump.