Fix error with DISTINCT on a datatype without btree opclass.
Without this patch:

    select distinct x from xidtab;
    ERROR:  could not identify an ordering operator for type xid
    LINE 1: select distinct x from xidtab;
                            ^
    HINT:  Use an explicit ordering operator or modify the query.

There were two similar but distinct issues:

1. transformDistinctToGroupBy() called addTargetToSortList(), even though
   the expression might not have ordering operators. There's another
   function, addTargetToGroupList(), for that. Use that.

2. make_distribution_exprs_for_groupclause() tried to create a PathKey on
   the expression, even if it didn't have ordering operators.

Also update a comment to explain why we do the DISTINCT -> GROUP BY
transformation. The comment said that it was because DISTINCT doesn't
support hashing, but that was fixed in PostgreSQL 8.4 already. However,
the planner still doesn't know how to perform the "pre-unique"
optimization for DISTINCT, so you still get better plans with GROUP BY.

This came up when working on the v12 merge. A new test was added to
upstream, 'hash_func', which contained a query like this involving
'relacl' datatype. But it's a pre-existing issue.

Fixes https://github.com/greenplum-db/gpdb/issues/9855
Reviewed-by: Pengzhou Tang <ptang@pivotal.io>