Harden resource group bypass test
Previously, this test depended on a very specific amount of memory being
"eaten" or consumed before it went OOM. ORCA using gpdb allocators
vs the legacy allocation (which is used in debug builds for leak
detection) changes the output. However, we have no way to detect or
change the output file depending on whether we're in debug build or not.

Now, instead of consuming the same amount of memory each time in the
test, we consume a small amount and ensure we don't go OOM. Then we
consume a larger amount and ensure we do go OOM. This hardens the test
while keeping the intent.