Stop cluster before changing system limits
During the behave tests, we change some of the system limits such as
memory and overcommit configurations. Previously this was changed while
the cluster was running, which caused the container to crash when
setting some of these parameters with ORCA in debug build. Now, we stop
the cluster, change the limits, and start the cluster.