Fix flaky uao_crash_compaction_row test
This test creates an AO table and inserts data on all segments, then
it deletes some data on seg0 and seg1 and do a vacuum on the AO
table. When doing vacuum, it suspends the QE in seg0 at starting
doing the post vacuum cleanup, then crash the seg0 and finally do
the post crash validation checks using gp_toolkit.__gp_aoseg(), this
function will check all aoseg info on all segments.

The VACUUM process on seg1 is in an uncertain state, it might have
finished the post cleanup which is expected or hasn't started yet,
so the aoseg info in seg1 is uncertain too.

To resolve this, this commit added a new injector on the point all
post vacuum cleanup are committed and validate the aoseg info after
the vacuum process on seg1 reached this point.