Account for transvalue in combine function in the mpool
We found that two-stage hashagg wasn't spilling early enough in the
final aggregate. This was because the transvalue wasn't being accounted
for in nodeAgg. This sometimes results in 200x under-counting. Since we
base our spilling decisions on the mpool's accounting, this commit is
necessary. This seems to be an oversight in commit 38d88155520 (9.6
merge commit). What happend in the merge is:
advance_combine_function() was never used in 6X, instead
invoke_agg_trans_func() was used and which used datumCopyWithMemManager().
When we started calling advance_combine_function() in the 9_6 merge
instead, we didn't adapt it to use datumCopyWithMemManager().

Since we also intend to backport upstream's disk-based hashagg
implementation into GP7, and possibly refactor/re-purpose/remove mpool
it is very much plausible that we will fall back to the upstream code
hunks around this change and rely on MemoryContext level accounting.
This is why we if-0-ed those.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>