Reconsider direct dispatch when Result node's contains gp_execution_segment() = <segid>
Greenplum may use the skill that setting Result plan's resconstantqual to
be an expression of gp_execution_segment() = <segid>. This skill is used
in to turn a general locus path to partitioned locus path in the function
set_append_path_locus. When using this skill, we should re-consider direct
dispatch info at the function of create_projection_plan to handle it correctly.

This commit fixes github issue: https://github.com/greenplum-db/gpdb/issues/9874.
For more details, refer to the issue.