Fix two subquery bugs.
1. If a join qual in the subquery contains references to the outer query,
it must be evaluated in the the same slice as the outer-query. Otherwise
the outer variables won't be available where they're evaluated. Teach
cdbpath_motion_for_join() to take that into account.

2. After fixing the first bug, one of the new test queries was still
failing with a runtime error:

SELECT (SELECT join_tab1.i - join_tab2.i
        FROM join_tab1, join_tab2 WHERE join_tab1.i = join_tab2.i and out1.i = out2.i
        LIMIT 1) as x
FROM join_tab1 out1, join_tab2 out2;
ERROR:  cannot execute squelched plan node of type: 203 (execProcnode.c:948)  (seg0 slice4 127.0.0.1:40000 pid=10725) (execProcnode.c:948)

If a Limit is used in a subquery, we better not squelch it if we expect the
subquery to be rescanned.

The first bug was introduced in the refactoring of subquery processing on
'master' branch, but the second bug is present in 6X_STABLE, too.

Reviewed-by: Zhenghua Lyu <zlv@pivotal.io>