Stop "normalizing" the join tree of every query.
Some of the GPDB-added subquery pull-up operations relied on the join
tree having specific form. But it's straightforward to teach those
operations to deal with arbitarary FromExprs and JoinExprs in the join
tree. Do that, and remove the join tree normalization step.

The reason I'm hacking on this right now is that on the 'merge' branch,
where we're working on merging with PostgreSQL v12, the normalization was
messing with the new SRF-in-targetlist processing. It probably would've
been easy to fix, but if we can get rid of the normalization step
altogether, that's even better.

Reviewed-by: Richard Guo <riguo@pivotal.io>
Reviewed-by: Yandong Yao <yyao@pivotal.io>