Bring back load balancing of single segment index for gang type GANGTYPE_SINGLETON_READER. (#9995)
It was removed during the slice refactor work. I found that when running test
isolation2/terminate_in_gang_creation.  This feature should be quite useful in
parallel OLTP load.

With this code change, test isolation2/terminate_in_gang_creation needs to be
modified to be deterministic. Also added tests in regression/dispatch for this
code change.  Changes in other tests are not relevant (just for cleanup).

Reviewed-by: Zhenghua Lyu <zlv@pivotal.io>