Remove gcc_env.sh and use system toolchain
Rather than relying on a specific compiler and toolchain setup in the
pipeline, we anticipate a valid compiler and toolchain is provided
either by setting of "CC" and "CXX" variables or modifying path and
LD_LIBRARY_PATH appropriately.

As part of this, we expect that necessary standard lib shared objects
are provided by the system or the build process uses a GCC/G++ version
that statically links them, which allows us to unvendor c++ stdlib for
Centos.

Co-authored-by: Tyler Ramer <tramer@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>