Revert "Eliminating alien nodes on master. (#9938)"
This commit is used to optimized the CPU cost on master
by removing the unnecessary ExecInitNode() on all the plan
nodes. But some works have to be done on master, e.g. pxf
use ExecInitNode to fetch meta data from external source only
on master. One solution is to add a blacklist for each of these
corner cases, but we need to evaluate it whether it is worth to
do that.

Currently ExecInitNode() is not the CPU bottleneck, Let's revert
this commit firstly and re-evalute the pros and cons in future.

This reverts commit 874939f47f64e472395a807ad28915b8e4d7f59e.