Move a query to a resource group
The syntax is: select pg_resgroup_move_query(pid, groupName);

It moves the query ('pid' is the process id of QD) to the resource group
'groupName'. This query will hang if the destination group doesn't have enough
resource (slot or memory) and it will be wake up when the resource become
available.

It cannot move a query that is in 'IDLE' state or the query is waiting on other
resource group.