Output CXXFLAGS together with CFLAGS in configure.
This makes it much more clear to the user (read: me) what configure has
settled on for C++ compilation flags. This prepares us for a forthcoming
changes which changes how we manipulate CXXFLAGS in configure. This is
written to *always* output CXXFLAGS even if we dis-configure C++ parts
(gpcloud, ORCA, etc) to match the unconditional AC_PROG_CXX. And even
without the immediately following change, this change has merits on its
own.