Remove useless materialize plannode in left tree of join
When handling join, if the inner path's locus is CdbLocusType_OuterQuery,
it also set the outer path's locus to CdbLocusType_OuterQuery in the
function cdbpath_motion_for_join. And later in cdbpath_create_motion_path,
it will add a materialize path over the motion path if the original path's
locus is CdbLocusType_OuterQuery. If the whole subquery can never be rescanned,
this materialize path is useless and will lead to performance lose. A typical
plan before this patch (from the case regress/expected/join.out) is:

  explain (verbose, costs off)
  select * from int4_tbl a,
    lateral (
      select * from int4_tbl b left join int8_tbl c on (b.f1 = q1 and a.f1 = q2)
    ) ss;
                               QUERY PLAN
  ------------------------------------------------------------------------
   Nested Loop
     Output: a.f1, b.f1, c.q1, c.q2
     ->  Materialize  **< this plannode is useless >**
           Output: a.f1
           ->  Gather Motion 3:1  (slice1; segments: 3)
                 Output: a.f1
                 ->  Seq Scan on public.int4_tbl a
                       Output: a.f1
     ->  Materialize
           Output: b.f1, c.q1, c.q2
           ->  Hash Right Join
                 Output: b.f1, c.q1, c.q2
                 Hash Cond: (c.q1 = b.f1)
                 ->  Result
                       Output: c.q1, c.q2
                       Filter: (a.f1 = c.q2)

There are two cases that we can do this safely:

  * not in SubPlan
  * not in lateral join's inner subquery

This commit add a flag in PlannerInfo->config to guide
if we can remove the useless Materialize plannode in join's left tree.