Close short lived connections
Due to refactor of dbconn and newer versions of pygresql, using
`with dbconn.connect() as conn` no longer attempts to close a
connection, even if it did prior. Instead, this syntax uses the
connection itself as context and, as noted in execSQL, overrides the
autocommit functionality of execSQL.

Therefore, close the connection manually to ensure that execSQL is
auto-commited, and the connection is closed.

Co-authored-by: Tyler Ramer <tramer@pivotal.io>
Co-authored-by: Jamie McAtamney <jmcatamney@pivotal.io>