Remove optimization to share Sort tapes directly in ShareInputScan.
Previously, ShareInputScan could co-opearate with a Sort node, to share
the final sort tape directly with other processes. Remove that support. If
a Sort node is shared, we now put a Materialize node on top of the Sort,
like all other nodes.

That is obviously less performant than sharing the sort tape directly.
However, I don't believe that is significant in practice. Firstly, if you
consider how a ShareInputScan is used, having a Sort below the
ShareInputScan should be rare. A ShareInputScan is used to implement CTEs,
and in order to have a Sort node just below the ShareInputScan, you need
to have an ORDER BY in the CTE. For example (from the 'sisc_sort_spill'
test):

    select avg(i3) from (
      with ctesisc as (select * from testsisc order by i2)
      select t1.i3, t2.i2
      from ctesisc as t1, ctesisc as t2
      where t1.i1 = t2.i2
    ) foo;

However, in a query like that, the ORDER BY is actually useless; the
order is not guaranteed to be preserved. In fact, ORCA optimizes it away
completely.

Secondly, even if you have a query like that, I don't think optimizing
away the Material is very significant. If the number of rows is small
enough to fit in memory, the Sort can be performed in memory, so you're
still writing it to disk only once, in the Material node. If it's large
enough to spill, the Material node will shield the Sort node from needing
to support random access, which enables the "on-the-fly" final merge
optimization in the tuplesort. So I believe you'll do roughly the same
amount of I/O in that case, too. One way to think about this is that the
final merge will be written out to the Material's tuplestore instead of
the tuplesort's file.

There is one drawback to that: the Material node won't be able to reuse
the disk space used by the sort tapes, as the final merge is performed,
so you'll momentarily need twice the disk space. I think that's
acceptable. If you don't like that, don't put superfluous ORDER BYs in
your WITH queries.

Reviewed-by: Georgios Kokolatos <gkokolatos@pivotal.io>